:bug: Corrigir a quantidade de casas decimais para versoes do Android abaixo de N.

+++ app/src/main/java/com/br/minasfrango/util/ImpressoraUtil.java
@@ -269,7 +269,7 @@ public class ImpressoraUtil {
                             + " KG"
                             + "{br}");
         } else {
-            int pesoTotal = 0;
+            Double pesoTotal = 0.0;
             for (ItemPedido itemPedido : pedido.getItens()) {
                 pesoTotal += itemPedido.getQuantidade();
             }
